@@ -1,4 +1,4 @@
-class AddCodeCoverageAnalysisBelongsToToReports < ActiveRecord::Migration
+class AddCodeCoverageAnalysisBelongsToToReports < ActiveRecord::Migration[4.2]
   def change
     add_reference :reports, :code_coverage_analysis, index: true, foreign_key: true
   end
