@@ -1,4 +1,4 @@
-class CreateGithubAccounts < ActiveRecord::Migration
+class CreateGithubAccounts < ActiveRecord::Migration[4.2]
   def change
     create_table :github_accounts do |t|
       t.string :name
