@@ -1,3 +1,10 @@
-//= require jquery
-//= require jquery_ujs
-//= require_tree .
+//= require jquery3
+//= require rails-ujs
+//= require materialize
+//= require turbolinks
+
+$( document ).on('turbolinks:load', function() {
+  $(".dropdown-trigger").dropdown();
+
+
+})
