@@ -6,8 +6,8 @@ Rails.application.configure do
     address:              "smtp.gmail.com",
     port:                 587,
     domain:               "gorgeouscode.com",
-    user_name:            ENV["EMAIL"],
-    password:             ENV["PASSWORD"],
+    user_name:            Rails.application.secrets.smtp_username,
+    password:             Rails.application.secrets.smtp_password,
     authentication:       "plain",
     enable_starttls_auto: true
   }
